refactor: Standardize naming convention with suffix `Provider` and capitalcase filenames for tests
