+++ 03_Modules/OcppRouter/package.json
@@ -9,7 +9,7 @@
   ],
   "scripts": {
     "prepublish": "npx eslint",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest --config jest.config.ts"
   },
   "keywords": [
     "ocpp",
